requirements.txt
Flask
requests
openai
google.generativeai